viewmodel ajustes cambio preguntas a int y cambio a accion de su set, cambio a como aparece el nuero de preguntas en el slider

+++ TriviaApp/app/src/main/java/com/example/triviaapp/componentes/ComponenteSlider.kt
@@ -34,7 +34,7 @@ fun ComponenteSlider(
         modifier = Modifier
             .padding(10.dp)
     ) {
-        Text(text= "Número de preguntas: " + datos.estadoSlider,
+        Text(text= "Número de preguntas: " + datos.estadoSlider.toInt(),
             color = MaterialTheme.colorScheme.onBackground
         )
         Slider(
